Added getBiggestSpender and refactored mostPopularCountry

+++ chinook/src/main/java/com/example/chinook/domain/models/CustomerSpender.java
@@ -1,4 +1,4 @@
 package com.example.chinook.domain.models;
 
-public record CustomerSpender(int customerId, String firstName, String lastName, int totalTransactions, int totalSpent) {
+public record CustomerSpender(int customerId, String firstName, String lastName, int totalTransactions, double totalSpent) {
 }
